pyproject.toml
[project]
name = "main-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas<3",
    "pillow<12",
    "plotly>=6.6.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.35.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]
